fix stf

+++ flood_forecast/pytorch_training.py
@@ -424,8 +424,8 @@ def compute_validation(validation_loader: DataLoader,
                     pred_len = model.pred_len
                     label_len = model.label_len
                     dec_inp = torch.zeros_like(trg_b[:, -pred_len:, :]).double()
-                    dec_inp = torch.cat([trg_b[:, :label_len, :], dec_inp], dim=1).double().to(device)
-                    output = model(src[0].double(), src[1].double(), dec_inp, targ[0].double())
+                    dec_inp = torch.cat([trg_b[:, :label_len, :], dec_inp], dim=1).float().to(device)
+                    output = model(src[0], src[1], dec_inp, targ[0])
 
                 else:
                     output = simple_decode(model=model,
